changed: make SIMAD a private template
to limit rebuild cost and downstream build times
